remove error of rtol determination

+++ src/hamiltonian/mod.rs
@@ -109,13 +109,13 @@ impl PeriodicNextNearestXXZ{
 }
 
 
-pub fn prepare_energy_map<V>(index : V, energies : &Array1<f64>, unit : f64) -> FnvHashMap<i32, Vec<(V, usize)>>
+pub fn prepare_energy_map<V>(index : V, energies : &Array1<f64>, unit : f64) -> FnvHashMap<i128, Vec<(V, usize)>>
     where V : EigenValue + Clone{
     // Prepare hashmap which will be used for degeneracy check
 
-    let mut energy_map : FnvHashMap<i32, Vec<(V, usize)>> = FnvHashMap::default();
+    let mut energy_map : FnvHashMap<i128, Vec<(V, usize)>> = FnvHashMap::default();
     for (idx, &e) in energies.iter().enumerate(){
-        let k = (e / unit).floor() as i32;
+        let k = (e / unit).floor() as i128;
         match energy_map.get_mut(&k){
             None => {
                 match energy_map.get_mut(&(k -1)){
@@ -143,13 +143,13 @@ pub fn prepare_energy_map<V>(index : V, energies : &Array1<f64>, unit : f64) -> 
     return energy_map;
 }
 
-pub fn count_degeneracy_from<V>(energy_map : &mut FnvHashMap<i32, Vec<(V, usize)>>, index : V, energies : &Array1<f64>, unit : f64) -> bool
+pub fn count_degeneracy_from<V>(energy_map : &mut FnvHashMap<i128, Vec<(V, usize)>>, index : V, energies : &Array1<f64>, unit : f64) -> bool
     where V : EigenValue + Clone  + std::fmt::Debug{
     // Store degeneracy information only for value already in energy_map
 
     let mut t = false;
     for (idx, &e) in energies.iter().enumerate(){
-        let k = (e / unit).floor() as i32;
+        let k = (e / unit).floor() as i128;
         match energy_map.get_mut(&k){
             None => {
                 match energy_map.get_mut(&(k - 1)){
@@ -177,7 +177,7 @@ pub fn count_degeneracy_from<V>(energy_map : &mut FnvHashMap<i32, Vec<(V, usize)
     return t;
 }
 
-pub fn degeneracy_pair<V>(length : usize, energy_map : &FnvHashMap<i32, Vec<(V, usize)>>) -> Result<(Array1<f64>, FnvHashMap<V, Vec<(usize, usize)>>), Error>
+pub fn degeneracy_pair<V>(length : usize, energy_map : &FnvHashMap<i128, Vec<(V, usize)>>) -> Result<(Array1<f64>, FnvHashMap<V, Vec<(usize, usize)>>), Error>
     where V : EigenValue + Hash + Eq{
     let mut pair_map : FnvHashMap<V, Vec<(usize, usize)>> = FnvHashMap::default();
     let mut counts : Array1<f64> = Array1::ones(length);
@@ -207,7 +207,7 @@ pub fn degeneracy_pair<V>(length : usize, energy_map : &FnvHashMap<i32, Vec<(V, 
     return Ok((counts, pair_map));
 }
 
-pub fn degeneracy_triple<V>(length : usize, energy_map : &FnvHashMap<i32, Vec<(V, usize)>>) -> Result<(Array1<f64>, FnvHashMap<V, Vec<(usize, usize, usize)>>), Error>
+pub fn degeneracy_triple<V>(length : usize, energy_map : &FnvHashMap<i128, Vec<(V, usize)>>) -> Result<(Array1<f64>, FnvHashMap<V, Vec<(usize, usize, usize)>>), Error>
     where V : EigenValue + Hash + Eq{
     let mut pair_map : FnvHashMap<V, Vec<(usize, usize, usize)>> = FnvHashMap::default();
     let mut counts : Array1<f64> = Array1::ones(length);
